chore: ui refactor resume.tsx to hero.tsx
Logically abstracted out and renamed `Hero` component to its own folder, and corrected / renamed component filename.
